refactor(frontend): extract TransactionsPage to src/pages/ (Phase 3)
Part of the Phase 3 migration. Route shrinks from 448 to 24 lines
(search-param schema + thin wiring only).

The filters drawer still uses useDisclosure + a prop-controlled
Mantine Drawer at the bottom of the mobile layout; converting it to
renderDrawer is tracked with a TODO for Phase 5.

Replaces Route.useSearch() / Route.useNavigate() in-page with the
from-qualified useSearch({ from }) / useNavigate({ from }) pattern
already used in TextSearch.tsx.
